Change chainId type from Int to BigInteger

+++ app/src/test/java/io/gnosis/safe/ui/assets/collectibles/CollectiblesViewModelTest.kt
@@ -59,7 +59,7 @@ class CollectiblesViewModelTest {
     fun `load - should emit collectibles view data list`() {
         val stateObserver = TestLiveDataObserver<BaseStateViewModel.State>()
         val chain = Chain(
-            BuildConfig.CHAIN_ID,
+            BuildConfig.CHAIN_ID.toBigInteger(),
             BuildConfig.BLOCKCHAIN_NAME,
             BuildConfig.CHAIN_TEXT_COLOR,
             BuildConfig.CHAIN_BACKGROUND_COLOR
@@ -147,5 +147,4 @@ class CollectiblesViewModelTest {
             null
         )
     )
-
 }
