+++ data/src/main/java/io/gnosis/data/backend/GatewayApi.kt
@@ -11,6 +11,7 @@ import io.gnosis.data.models.transaction.TransactionConfirmationRequest
 import io.gnosis.data.models.transaction.TransactionDetails
 import io.gnosis.data.models.transaction.TxListEntry
 import retrofit2.http.*
+import java.math.BigInteger
 import java.util.*
 
 interface GatewayApi {
@@ -20,53 +21,53 @@ interface GatewayApi {
 
     @GET("/v1/chains/{chainId}/safes/{address}")
     suspend fun getSafeInfo(
-        @Path("chainId") chainId: Int,
+        @Path("chainId") chainId: BigInteger,
         @Path("address") address: String
     ): SafeInfo
 
     @GET("/v1/chains/{chainId}/safes/{address}/balances/{fiat}")
     suspend fun loadBalances(
-        @Path("chainId") chainId: Int,
+        @Path("chainId") chainId: BigInteger,
         @Path("address") address: String,
         @Path("fiat") fiat: String = "usd"
     ): CoinBalances
 
     @GET("/v1/chains/{chainId}/transactions/{transactionId}")
     suspend fun loadTransactionDetails(
-        @Path("chainId") chainId: Int,
+        @Path("chainId") chainId: BigInteger,
         @Path("transactionId") transactionId: String
     ): TransactionDetails
 
     @POST("/v1/chains/{chainId}/transactions/{safeTxHash}/confirmations")
     suspend fun submitConfirmation(
-        @Path("chainId") chainId: Int,
+        @Path("chainId") chainId: BigInteger,
         @Path("safeTxHash") safeTxHash: String,
         @Body txConfirmationRequest: TransactionConfirmationRequest
     ): TransactionDetails
 
     @POST("/v1/chains/{chainId}/transactions/{safeAddress}/propose")
     suspend fun proposeTransaction(
-        @Path("chainId") chainId: Int,
+        @Path("chainId") chainId: BigInteger,
         @Path("safeAddress") safeAddress: String,
         @Body multisigTransactionRequest: MultisigTransactionRequest
     )
 
     @GET("/v1/chains/{chainId}/safes/{safeAddress}/collectibles")
     suspend fun loadCollectibles(
-        @Path("chainId") chainId: Int,
+        @Path("chainId") chainId: BigInteger,
         @Path("safeAddress") safeAddress: String
     ): List<Collectible>
 
     @GET("/v1/chains/{chainId}/safes/{address}/transactions/history")
     suspend fun loadTransactionsHistory(
-        @Path("chainId") chainId: Int,
+        @Path("chainId") chainId: BigInteger,
         @Path("address") address: String,
         @Query("timezone_offset") timezoneOffset: Int = TimeZone.getDefault().getOffset(Date().time)
     ): Page<TxListEntry>
 
     @GET("/v1/chains/{chainId}/safes/{address}/transactions/queued")
     suspend fun loadTransactionsQueue(
-        @Path("chainId") chainId: Int,
+        @Path("chainId") chainId: BigInteger,
         @Path("address") address: String,
         @Query("timezone_offset") timezoneOffset: Int = TimeZone.getDefault().getOffset(Date().time)
     ): Page<TxListEntry>
